+++ bopep/__init__.py
@@ -1,7 +1,7 @@
-from bopep.bayesian_optimization.optimization import BoPep
+from bopep.search.optimization import BoPep
 from bopep.scoring.scorer import Scorer
 from bopep.embedding.embedder import Embedder
 from bopep.docking.docker import Docker
 from bopep.scoring.is_peptide_in_binding_site import get_binding_site
 from bopep.scoring.scores_to_objective import bopep_objective, benchmark_objective
-from bopep.synthesis.synthesizer import Synthesizer
+from bopep.design.borf import Borf
