+++ bopep/design/__init__.py
@@ -1,5 +1,5 @@
 from bopep.design.mpnn_fastrelax import MPNNFastRelax
 from bopep.design.diffusion import RFDiffusion
-from bopep.design.synthesizer import borf
+from bopep.design.borf import Borf
 
-__all__ = ["MPNNFastRelax", "RFDiffusion", "borf"]
+__all__ = ["MPNNFastRelax", "RFDiffusion", "Borf"]
