+++ bopep/search/utils.py
@@ -3,7 +3,7 @@ import subprocess
 from typing import Any, Dict, List, Optional, Union
 import torch
 
-from bopep.bayesian_optimization.acquisition_functions import available_acquistion_functions
+from bopep.search.acquisition_functions import available_acquistion_functions
 
 def _save_model(save_path: str, model, surrogate_model_kwargs: dict, best_hyperparams: dict):
         """
